refactor(summary-notification): get_logger / get_settings を lru_cache ベースに統一
asset-collection と同様に、引数注入パターンを廃止して lru_cache ベースの
ファクトリ関数に統一する。get_logger のインポートパスを src.config に集約する。
