+++ lambda/summary-notification/src/config/__init__.py
@@ -1,0 +1,5 @@
+from shared.config.base_settings import get_logger
+
+from .settings import EnvSettings, get_settings
+
+__all__ = ["get_logger", "EnvSettings", "get_settings"]
