+++ lambda/summary-notification/src/config/settings.py
@@ -1,4 +1,6 @@
-from shared.config.base_settings import BaseEnvSettings, get_logger  # noqa: F401
+from functools import lru_cache
+
+from shared.config.base_settings import BaseEnvSettings
 
 
 class EnvSettings(BaseEnvSettings):
@@ -9,15 +11,6 @@ class EnvSettings(BaseEnvSettings):
     spreadsheet_parameter_name: str
 
 
-def get_settings(settings_instance: EnvSettings | None = None) -> EnvSettings:
-    """設定インスタンスを取得する
-
-    Args:
-        settings_instance (EnvSettings | None, optional): 設定インスタンス. Defaults to None.
-
-    Returns:
-        EnvSettings: 設定インスタンス
-    """
-    if settings_instance is None:
-        settings_instance = EnvSettings()
-    return settings_instance
+@lru_cache()
+def get_settings() -> EnvSettings:
+    return EnvSettings()  # pyright: ignore[reportCallIssue]
