+++ lambda/summary-notification/src/handler.py
@@ -2,7 +2,7 @@ from aws_lambda_powertools.utilities.typing import LambdaContext
 from shared.domain.financial_asset_repository import IFinancialAssetRepository
 
 from src.application import INotifier, INotifyWeeklySummaryUseCase, NotifyWeeklySummaryUseCase
-from src.config.settings import EnvSettings, get_logger, get_settings
+from src.config import EnvSettings, get_logger, get_settings
 from src.infrastructure import (
     GoogleSheetFinancialAssetRepository,
     LineNotifier,
